requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
azure-storage-blob>=12.19.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic[email]>=2.9.0
authlib>=1.2.1
httpx>=0.25.2
cryptography>=41.0.7
pyjwt[crypto]>=2.8.0
posthog>=3.0.0

